feat: QA 건의사항 반영
