+++ frontend/src/components/host/SectionCard/index.tsx
@@ -1,7 +1,6 @@
-import ToggleSwitch from '@/components/common/ToggleSwitch';
-import useToggleSwitch from '@/components/common/ToggleSwitch/useToggleSwitch';
-import SectionCardDefault from '@/components/host/SectionCardDefault';
-import SectionCardDetail from '@/components/host/SectionCardDetail';
+import TaskBox from '../TaskBox';
+import useSectionCardDefault from './useSectionCardDefault';
+import { BiNews, BiX } from 'react-icons/bi';
 
 import { SectionType } from '@/types';
 
@@ -13,20 +12,29 @@ interface SectionCardProps {
 }
 
 const SectionCard: React.FC<SectionCardProps> = ({ section, sectionIndex }) => {
-  const { toggle, onClickToggle } = useToggleSwitch();
-
-  const toggleLeft = !toggle;
+  const { onChange, onClickDelete, onClickCreate, onClickSectionDetail, hasSectionDetailInfo } =
+    useSectionCardDefault(sectionIndex);
 
   return (
     <div css={styles.container}>
-      <div css={styles.switchWrapper}>
-        <ToggleSwitch toggle={toggle} onClickSwitch={onClickToggle} left="기본" right="상세정보" />
+      <BiX css={styles.deleteButton} size={30} onClick={onClickDelete} />
+      <div css={styles.titleWrapper}>
+        <input
+          css={styles.input}
+          placeholder="새 구역"
+          defaultValue={section.name}
+          maxLength={10}
+          onChange={onChange}
+          required
+        />
+        <BiNews css={styles.detailButton(hasSectionDetailInfo())} size={26} onClick={onClickSectionDetail} />
       </div>
-      {toggleLeft ? (
-        <SectionCardDefault section={section} sectionIndex={sectionIndex} />
-      ) : (
-        <SectionCardDetail section={section} sectionIndex={sectionIndex} />
-      )}
+      {section.tasks.map((task, taskIndex) => (
+        <TaskBox task={task} taskIndex={taskIndex} sectionIndex={sectionIndex} key={task.id} />
+      ))}
+      <button css={styles.newTaskButton} onClick={onClickCreate} type="button">
+        + 새 작업 추가
+      </button>
     </div>
   );
 };
