+++ frontend/src/components/host/SectionCard/styles.ts
@@ -1,5 +1,6 @@
 import { css } from '@emotion/react';
 
+import animation from '@/styles/animation';
 import theme from '@/styles/theme';
 
 const container = css`
@@ -9,12 +10,14 @@ const container = css`
   max-width: 480px;
   height: 360px;
   overflow-y: scroll;
-  padding: 16px 32px;
+  padding: 32px;
   background-color: ${theme.colors.white};
   box-shadow: 2px 2px 2px 2px ${theme.colors.shadow20};
   border-radius: 8px;
+  position: relative;
+
   ::-webkit-scrollbar {
-    width: 0.4em;
+    display: none;
   }
 
   ::-webkit-scrollbar-thumb {
@@ -35,16 +38,60 @@ const container = css`
   }
 `;
 
-const switchWrapper = css`
+const titleWrapper = css`
   display: flex;
-  justify-content: center;
-  width: 100%;
-  margin-bottom: 24px;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 20px;
+  margin-bottom: 8px;
+  padding-bottom: 16px;
+  border-bottom: 2px solid ${theme.colors.shadow20};
 `;
 
-const styles = {
-  container,
-  switchWrapper,
-};
+const deleteButton = css`
+  top: 8px;
+  right: 8px;
+  position: absolute;
+  margin: 0 4px;
+  color: ${theme.colors.gray800};
+  cursor: pointer;
+`;
+
+const detailButton = (hasSectionDetailInfo: boolean) => css`
+  margin: 0 4px;
+  cursor: pointer;
+  color: ${hasSectionDetailInfo ? theme.colors.green : theme.colors.gray400};
+  :hover {
+    animation: ${animation.shake} 2s infinite linear alternate;
+  }
+`;
+
+const newTaskButton = css`
+  align-self: center;
+  margin: 8px 0;
+  font-size: 12px;
+  padding: 8px 16px;
+  border-radius: 24px;
+  background-color: ${theme.colors.shadow10};
+  :hover {
+    background-color: ${theme.colors.shadow20};
+  }
+`;
+
+const input = css`
+  font-size: 18px;
+  line-height: 38px;
+  border: 1px solid ${theme.colors.shadow30};
+  border-radius: 12px;
+  padding: 0 16px;
+  background-color: ${theme.colors.white};
+  width: 80%;
+
+  :focus {
+    outline: 2px solid ${theme.colors.primary};
+  }
+`;
+
+const styles = { container, titleWrapper, input, detailButton, deleteButton, newTaskButton };
 
 export default styles;
