+++ frontend/src/components/host/TaskBox/index.tsx
@@ -1,7 +1,5 @@
 import useTaskBox from './useTaskBox';
-import { BiPencil, BiTrash } from 'react-icons/bi';
-
-import Button from '@/components/common/Button';
+import { BiMinus, BiNews, BiXCircle } from 'react-icons/bi';
 
 import { TaskType } from '@/types';
 
@@ -14,19 +12,25 @@ interface TaskBoxProps {
 }
 
 const TaskBox: React.FC<TaskBoxProps> = ({ task, taskIndex, sectionIndex }) => {
-  const { onChangeInput, onClickDeleteButton } = useTaskBox(taskIndex, sectionIndex);
+  const { onChangeInput, onClickDeleteButton, onClickTaskDetail, hasTaskDetailInfo } = useTaskBox(
+    taskIndex,
+    sectionIndex
+  );
 
   return (
     <div css={styles.taskBox}>
-      <input
-        css={styles.input}
-        placeholder="새 작업"
-        defaultValue={task.name}
-        maxLength={10}
-        onChange={onChangeInput}
-        required
-      />
-      <BiTrash css={styles.trash} size={22} onClick={onClickDeleteButton} />
+      <div>
+        <BiMinus css={styles.deleteButton} size={18} onClick={onClickDeleteButton} />
+        <input
+          css={styles.input}
+          placeholder="새 작업"
+          defaultValue={task.name}
+          maxLength={10}
+          onChange={onChangeInput}
+          required
+        />
+      </div>
+      <BiNews css={styles.detailButton(hasTaskDetailInfo())} size={26} onClick={onClickTaskDetail} />
     </div>
   );
 };
